Deprecated ChildLogin and RoleSelect. Refactored ParentLogin into UserLogin. Both Child and Parent accounts can log in from the same page.
