+++ app/src/main/java/com/taskmaster/appui/Page/ProgressionPage.java
@@ -38,6 +38,7 @@ import com.google.firebase.firestore.QuerySnapshot;
 import com.taskmaster.appui.Page.Login.Splash;
 import com.taskmaster.appui.Page.Main.QuestManagement;
 import com.taskmaster.appui.R;
+import com.taskmaster.appui.Services.NavUtil;
 
 import java.util.ArrayList;
 import java.util.List;
