+++ app/src/main/res/layout/parent_login.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Page.Login.ParentLogin">
+    tools:context=".Page.Login.UserLogin">
 
     <ImageView
         android:id="@+id/imageView6"
@@ -54,9 +54,10 @@
         android:alpha="0.6"
         app:layout_constraintBottom_toBottomOf="@+id/imageView7"
         app:layout_constraintEnd_toEndOf="@+id/imageView7"
+        app:layout_constraintHorizontal_bias="0.485"
         app:layout_constraintStart_toStartOf="@+id/imageView7"
         app:layout_constraintTop_toTopOf="@+id/imageView7"
-        app:layout_constraintVertical_bias="0.475"
+        app:layout_constraintVertical_bias="0.46"
         app:srcCompat="@drawable/rectangle_rounded" />
 
     <ImageView
@@ -85,14 +86,14 @@
         android:id="@+id/textView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="55dp"
+        android:layout_marginTop="48dp"
         android:fontFamily="@font/eb_garamond_semibold"
         android:text="Log In"
         android:textAlignment="center"
         android:textColor="#3F3F3F"
         android:textSize="24sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="@+id/imageView10"
+        app:layout_constraintStart_toStartOf="@+id/imageView10"
         app:layout_constraintTop_toTopOf="@+id/imageView10" />
 
     <TextView
@@ -118,26 +119,23 @@
         android:id="@+id/signupTextView"
         android:layout_width="252dp"
         android:layout_height="19dp"
-        android:layout_marginStart="44dp"
-        android:layout_marginTop="433dp"
         android:layout_marginEnd="44dp"
-        android:layout_marginBottom="26dp"
+        android:layout_marginBottom="24dp"
         android:fontFamily="@font/eb_garamond_semibold"
         android:text="Need an account? Sign Up!"
         android:textAlignment="center"
         android:textColor="#3F3F3F"
         android:textSize="12sp"
         app:layout_constraintBottom_toBottomOf="@+id/imageView10"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/imageView10" />
+        app:layout_constraintEnd_toEndOf="@+id/imageView10"
+        app:layout_constraintHorizontal_bias="1.0"
+        app:layout_constraintStart_toStartOf="@+id/imageView10" />
 
     <EditText
         android:id="@+id/emailBox2"
         android:layout_width="255dp"
         android:layout_height="55dp"
-        android:layout_marginTop="127dp"
-        android:layout_marginBottom="315dp"
+        android:layout_marginTop="24dp"
         android:background="@drawable/textbox_rounded"
         android:ems="10"
         android:fontFamily="@font/eb_garamond_semibold"
@@ -147,17 +145,15 @@
         android:textAlignment="center"
         android:textColor="#3F3F3F"
         android:textSize="24sp"
-        app:layout_constraintBottom_toBottomOf="@+id/imageView10"
         app:layout_constraintEnd_toEndOf="@+id/imageView10"
         app:layout_constraintStart_toStartOf="@+id/imageView10"
-        app:layout_constraintTop_toTopOf="@+id/imageView10" />
+        app:layout_constraintTop_toBottomOf="@+id/textView" />
 
     <EditText
         android:id="@+id/passwordBox2"
         android:layout_width="255dp"
         android:layout_height="55dp"
-        android:layout_marginTop="197dp"
-        android:layout_marginBottom="226dp"
+        android:layout_marginTop="24dp"
         android:background="@drawable/textbox_rounded"
         android:ems="10"
         android:fontFamily="@font/eb_garamond_semibold"
@@ -167,17 +163,15 @@
         android:textAlignment="center"
         android:textColor="#3F3F3F"
         android:textSize="24sp"
-        app:layout_constraintBottom_toBottomOf="@+id/imageView10"
         app:layout_constraintEnd_toEndOf="@+id/imageView10"
         app:layout_constraintStart_toStartOf="@+id/imageView10"
-        app:layout_constraintTop_toTopOf="@+id/imageView10" />
+        app:layout_constraintTop_toBottomOf="@+id/emailBox2" />
 
     <androidx.appcompat.widget.AppCompatButton
         android:id="@+id/confirmButton2"
         android:layout_width="252dp"
         android:layout_height="55dp"
-        android:layout_marginTop="311dp"
-        android:layout_marginBottom="112dp"
+        android:layout_marginTop="48dp"
         android:background="@drawable/button_shadowed_states"
         android:fontFamily="@font/eb_garamond_semibold"
         android:stateListAnimator="@null"
@@ -186,10 +180,9 @@
         android:textAllCaps="false"
         android:textColor="#3F3F3F"
         android:textSize="24sp"
-        app:layout_constraintBottom_toBottomOf="@+id/imageView10"
         app:layout_constraintEnd_toEndOf="@+id/imageView10"
         app:layout_constraintStart_toStartOf="@+id/imageView10"
-        app:layout_constraintTop_toTopOf="@+id/imageView10"
+        app:layout_constraintTop_toBottomOf="@+id/passwordBox2"
         app:strokeColor="@color/black"
         app:strokeWidth="1dp" />
 
@@ -197,12 +190,12 @@
         android:id="@+id/imageView15"
         android:layout_width="233dp"
         android:layout_height="wrap_content"
-        android:layout_marginTop="417dp"
-        android:layout_marginBottom="60dp"
-        app:layout_constraintBottom_toBottomOf="@+id/imageView10"
+        android:layout_marginTop="192dp"
+        app:layout_constraintBottom_toTopOf="@+id/signupTextView"
         app:layout_constraintEnd_toEndOf="@+id/imageView10"
+        app:layout_constraintHorizontal_bias="0.514"
         app:layout_constraintStart_toStartOf="@+id/imageView10"
-        app:layout_constraintTop_toTopOf="@+id/imageView10"
+        app:layout_constraintVertical_bias="0.0"
         app:srcCompat="@drawable/line_1" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
